final1/2

+++ ShopCore/SM.Domain/OrderAgg/OrderDetailModel.cs
@@ -5,7 +5,7 @@ namespace SM.Domain.OrderAgg
     public class OrderDetailModel
     {
         protected OrderDetailModel() { }
-        public OrderDetailModel(int discountRate, int count, double unitPrice, int productId)
+        public OrderDetailModel(int discountRate, int count, double unitPrice, long productId)
         {
             DiscountRate = discountRate;
             Count = count;
@@ -23,7 +23,7 @@ namespace SM.Domain.OrderAgg
         public long OrderId { get;private set; }
         public OrderModel order { get;private set; }
 
-        public int ProductId { get; private set; }
+        public long ProductId { get; private set; }
        // public  ProductModel product { get; private set; }
     }
 }
